InterfaceWrapper.java -> Added a InterfaceWrapper.showErrorWindow method. It is used to show a warning/error message to the user. Purely informational.
Menu.java -> Added a change to layout null on ContentPanel instance, in case for some reason in other places it was easier to use another layout. Also added a new font size and bolded it to the menu name, to be more proeminent.

Main.java -> made some refactoring to accomodate assignature change in Game class. Also added case 1 (New Game) to the game management function handler (Main.gameManeg).

Game.java -> Changed the class signature from players being stored in a list to a simple number (was list because of 2x2, 1x1, 1, 2 2x2 1x1x1x1 and other formats). Created Game.createGameGUI, a function designed to interact with the GUI in order to receive the necessary data to create a new Game object.